day9 not yet there - opcode 3 not working

+++ src/day2/AdvancedCleverIntcodeProgram.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 class AdvancedCleverIntcodeProgram extends CleverIntcodeProgram {
 
-    AdvancedCleverIntcodeProgram(List<Integer> program, int noun, int verb) {
+    AdvancedCleverIntcodeProgram(List<Long> program, long noun, long verb) {
         this.program.clear();
         this.inputProgram.addAll(program);
         this.inputProgram.set(1, noun);
         this.inputProgram.set(2, verb);
     }
-    int getOutput(){
+    long getOutput(){
         return program.get(0);
     }
 }
